+++ src/kr_book_to_audio/__init__.py
@@ -1,2 +1,2 @@
 """KR Book To Audio public package."""
-__version__ = '2.3.2'
+__version__ = '2.3.3'